requirements.txt
streamlit
pandas
plotly
scikit-learn
pdfplumber
docx2txt
sentence-transformers
torch